requirements.txt
requests==2.32.3
openai==1.58.1
python-dotenv==1.1.1
pandas==2.3.3
matplotlib==3.10.6
mplfinance==0.12.10b0
python-dateutil==2.9.0
pytz==2024.1
coinbase-advanced-py>=1.2.0
